Bug 27005: (follow-up) Enable the exclusion of columns from export
This patch copies code from the staff interface which allows us to add a
class to table columns which should not be included in print or export
operations. The hidden column which facilitates the "Checkouts" and
"On-site checkouts" tabs can now be hidden in prints and exports.

To test, apply the patch and follow the previous test plan. When testing
the export and print buttons, confirm that the column with
"standard_checkout" data is not included.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>