Bug 27605: Add floating toolbar to patron search page
This patch modifies the patron search results page so that the toolbar
with results-related controls "floats" when the user scrolls down. This
gives access to the controls for adding patrons to a list and for
merging patrons.

Other templates are modified to incorporate a change to the CSS which
changes "#searchheader" to ".searchheader," enabling multiple instances
of a <div> styled with the .searchheader class.

To test, apply the patch and rebuild the staff interface CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

 - Go to Patrons in the staff interface.
 - Perform a patron search which will return many results.
 - Scroll down the page far enough to trigger the floating toolbar.
 - Confirm that the toolbar is positioned correctly when paging through
   results.
 - Confirm that the patron list and merge controls still work correctly.
 - Confirm that other pages which used the "#searchheader" id are still
   styled correctly:
   - Catalog search results
   - List contents
   - Patron list contents
   - Add orders from MARC file

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>