Bug 27628: Fix minor HTML markup errors in OPAC search results templates
This patch makes a couple of minor corrections to templates related to
OPAC search results: An errant quote, and invalidly repeating ids.

The #login4tags id is changed to a class, and CSS related to this id
(which was not used) is removed. For consistency, the id is changed to a
class on the detail page as well.

To test, apply the patch and rebuild the OPAC CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

 - Perform a search in the OPAC. Confirm that the "Log in to add tags"
   link still looks correct at various browser widths.
 - Confirm that it works to trigger the login modal.
 - Validate the source of the page. There should be no errors.

 - View the detail page for a bibliographic record.
 - The "Log in to add tags" link should look correct and work correctly.

Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>