Bug 27633: (bug 15436 follow-up) Fix duplicate of 440$v at the OPAC
Only display a in link, not v that is displayed later.

It has been fixed in different other places by
  commit 81f71d45fe4fee92596ea8d8edb57c602490e9b1
  Bug 15436: Use semicolon between series name and volume

but missed one occurrence.

Test plan:
Add a 440$a and $v and confirm that the display at the OPAC is now the
same as the staff interface.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
(cherry picked from commit b3c58aafcf249efdfe91b00d4da9c66472c4a141)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>