Bug 27497: Display Koha version in staff interface home page
Koha current version is very useful in staff interface.
You can find it in 'About' page but home page is the first page you
see. Display it in the lower right-hand side of the page, after
'IntranetmainUserblock' content.

Note that it is not a vertical floating text, it whould conflict with
languages floating bar.

Test plan :
1) Apply patch and update staff client CSS :
https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client
2) Look at staff interface home page and look at lower right-hand side
   of the page

Signed-off-by: Michael Adamyk <madamyk@ckls.org>

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>