Bug 18912: Display more info when using itemBarcodeFallbackSearch
Using itemBarcodeFallbackSearch entering a barcode on checkout will do a keyword
search and provide you with a list of items matching the search term for checkout.
This allows checking out items, where there is no barcode in the item (you still need
a dummy barcode in Koha).

At the moment the table of items to choose from will only chow the title and the dummy barcode.

In order to make it more useful and easier to identify the correct item,
we should show more information here:

- callnumber (952$o)
- serial enumeration (952$h)
- copy number (952$t)
- inventory number (952$i)

(description copier from comment 0)

Test plan:
- Turn on itemBarcodeFallbackSearch
(If you have ES you need bug 27597)
- Check an item out using "street" or a search that will return several
items
- Note that you have now more info about the items

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Victor Grousset/tuxayo <victor@tuxayo.net>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>