Bug 27728: Add a search box on OPAC Search history
This patch makes changes to the OPAC search history DataTables
configuration so that table controls appear, including search, copy,
CSV, and print.

The patch also makes some improvements to the checkbox interactions: The
date column is now a label for the checkbox, and the "Remove selected
searched" button is now disabled until a checkbox has been checked.

To test, apply the patch and log into the OPAC as a user with a search
history. If possible, past and current searches of both bibliographic
and authority records.

- On the search history page, confirm that the DataTables controls
  appear above each table: "Current" and "Previous" under both the
  "Catalog" and "Authority" tabs.
- All controls should work correctly and affect only the corresponding
  table: Search, Clear filter, Copy, CSV, and Print.
- Confirm that clicking a date in the date column checks the correct
  checkbox.
- Confirm that toggling a checkbox correctly enables and disables the
  "Remove selected searches" button as well as the "Select searches to:"
  menu at the top of the table.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>